delete now-unused vestiges of previous attempt
