+++ .vscode/launch.json
@@ -114,18 +114,6 @@
         "PYTHONPATH": "/Users/birch/git/pytorch_mps:${workspaceFolder}/src/diffusers/src:${workspaceFolder}/src/k-diffusion:${workspaceFolder}/src/lora:${workspaceFolder}/src:${env.PYTHONPATH}",
         "PYTORCH_ENABLE_MPS_FALLBACK": "1",
       }
-    },
-    {
-      "name": "Python: Seed Test 2",
-      "type": "python",
-      "request": "launch",
-      "program": "${workspaceFolder}/scripts/seed_test_2.py",
-      "console": "integratedTerminal",
-      "justMyCode": true,
-      "env": {
-        "PYTHONPATH": "/Users/birch/git/pytorch_mps:${workspaceFolder}/src/diffusers/src:${workspaceFolder}/src/k-diffusion:${workspaceFolder}/src/lora:${workspaceFolder}/src:${env.PYTHONPATH}",
-        "PYTORCH_ENABLE_MPS_FALLBACK": "1",
-      }
     }
   ]
 }
