+++ src/helpers/inference_spec/sample_spec.py
@@ -1,7 +1,8 @@
 from dataclasses import dataclass
 from .cond_spec import ConditionSpec
+from .latent_spec import LatentSpec
 
 @dataclass
 class SampleSpec:
-  seed: int
+  latent_spec: LatentSpec
   cond_spec: ConditionSpec
