--- src/helpers/inference_spec/sample_spec_2.py
@@ -1,8 +1,0 @@
-from dataclasses import dataclass
-from .cond_spec import ConditionSpec
-from .latent_spec import LatentSpec
-
-@dataclass
-class SampleSpec:
-  latent_spec: LatentSpec
-  cond_spec: ConditionSpec
